requirements.txt
# Core
#python>=3.11
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3

# OANDA
oandapyV20>=0.6.0

# Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
#ta-lib>=0.4.26

# Monitoring
python-telegram-bot>=20.0
psutil>=5.9.0

# Database (optional)
SQLAlchemy>=2.0.0
asyncpg>=0.27.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.2.0
flake8>=6.0.0
pre-commit>=3.3.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0